UI refactor for split pane view with allotment
