+++ html/src/components/functions/classNames.tsx
@@ -1,0 +1,3 @@
+export function classNames(...classes: string[]) {
+  return classes.filter(Boolean).join(' ')
+}
