+++ html/src/index.css
@@ -1,3 +1,5 @@
+@import "allotment/dist/style.css";
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
