+++ html/src/types/index.d.ts
@@ -1,0 +1,3 @@
+declare module "allotment" {
+  export * from "allotment/dist/types/src/index.d.ts";
+}
